feat: Use schema validator for list updates
