+++ test/unit/lists-handler.spec.ts
@@ -1345,118 +1345,6 @@ describe('when updating a list', () => {
     })
   })
 
-  describe('and the request body does not contain a valid JSON', () => {
-    beforeEach(() => {
-      jsonMock.mockRejectedValueOnce(new SyntaxError())
-    })
-
-    it('should return a response with a message saying that the body must be a parsable JSON and the 400 status code', () => {
-      return expect(updateListHandler({ components, verification, request, params })).resolves.toEqual({
-        status: StatusCode.BAD_REQUEST,
-        body: {
-          ok: false,
-          message: 'The body must contain a parsable JSON.'
-        }
-      })
-    })
-  })
-
-  describe('and the request body does not contain neither the name or the private properties', () => {
-    beforeEach(() => {
-      jsonMock.mockResolvedValueOnce({})
-    })
-
-    it('should return a response with a message saying that the name property is not correct and the 400 status code', () => {
-      return expect(updateListHandler({ components, verification, request, params })).resolves.toEqual({
-        status: StatusCode.BAD_REQUEST,
-        body: {
-          ok: false,
-          message: 'The body must contain at least one of the following properties: name or private.'
-        }
-      })
-    })
-  })
-
-  describe('and the request body does contains the name property but is not of type string', () => {
-    beforeEach(() => {
-      jsonMock.mockResolvedValueOnce({ name: 1 })
-    })
-
-    it('should return a response with a message saying that the name property is not correct and the 400 status code', () => {
-      return expect(updateListHandler({ components, verification, request, params })).resolves.toEqual({
-        status: StatusCode.BAD_REQUEST,
-        body: {
-          ok: false,
-          message: 'The property name is not of string type.'
-        }
-      })
-    })
-  })
-
-  describe('and the request body does contains the private property but is not of type boolean', () => {
-    beforeEach(() => {
-      jsonMock.mockResolvedValueOnce({ private: 'not a boolean' })
-    })
-
-    it('should return a response with a message saying that the private property is not correct and the 400 status code', () => {
-      return expect(updateListHandler({ components, verification, request, params })).resolves.toEqual({
-        status: StatusCode.BAD_REQUEST,
-        body: {
-          ok: false,
-          message: 'The property private is not of boolean type.'
-        }
-      })
-    })
-  })
-
-  describe('and the request body does contains the description property but is not of type string', () => {
-    beforeEach(() => {
-      jsonMock.mockResolvedValueOnce({ name, private: true, description: 1 })
-    })
-
-    it('should return a response with a message saying that the description property is not correct and the 400 status code', () => {
-      return expect(updateListHandler({ components, verification, request, params })).resolves.toEqual({
-        status: StatusCode.BAD_REQUEST,
-        body: {
-          ok: false,
-          message: 'The property description is not of string type.'
-        }
-      })
-    })
-  })
-
-  describe('and the request body does contains the name property but it longs more than 32 characters', () => {
-    beforeEach(() => {
-      jsonMock.mockResolvedValueOnce({ name: 'a'.repeat(33) })
-    })
-
-    it('should return a response with a message saying that the name property exceeds the 32 characters and the 400 status code', () => {
-      return expect(updateListHandler({ components, verification, request, params })).resolves.toEqual({
-        status: StatusCode.BAD_REQUEST,
-        body: {
-          ok: false,
-          message: 'The property name exceeds the 32 characters.'
-        }
-      })
-    })
-  })
-
-  describe('and the request body does contains a valid description but it longs more than 100 characters', () => {
-    beforeEach(() => {
-      jsonMock.mockResolvedValueOnce({ name: 'List'.repeat(8), description: 'desc'.repeat(26) })
-    })
-
-    it('should return a response with a message saying that the description property exceeds the 100 characters and the 400 status code', () => {
-      return expect(updateListHandler({ components, verification, request, params })).resolves.toEqual({
-        status: StatusCode.BAD_REQUEST,
-        body: {
-          ok: false,
-          message: 'The property description exceeds the 100 characters.'
-        }
-      })
-    })
-  })
-
   describe('and the process to update a list fails with a list not found error', () => {
     beforeEach(() => {
       jsonMock.mockResolvedValueOnce({ name })
